feat(db): add a state to the protocols
The `state` row contains an info about if the protocol should be
considered active. Inactive protocol should be impossible to assign to
the technicians.
Add a new table for planning the changes of the `state` row in the
`protocols` table. It works similarly to `users_change` table introduced
in f265058.

Refs: #5